+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-4925] Create new draft when editing a published implementation
             only if it has changes.
+- [TD-4997] Change search executions API sources list filter from 'AND' to 'OR'
 
 ## [4.47.0] 2022-07-05